Move LanguageClient to separate file
